requirements.txt
numpy==1.26.4
Pillow==10.2.0
tqdm==4.67.1
requests==2.33.1

torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

facenet-pytorch==2.6.0

opencv-python==4.9.0.80